Fixed broken addresses

+++ o6/testing/fuzz_test.c
@@ -1,8 +1,11 @@
-#include "../src/string.h"
+#include "../src/string_replace.c"
 #include <stddef.h>
 #include <stdint.h>
+#include <string.h>
 
-int LLVMFuzzerTestOneInput(char *data, size_t size) {
+int LLVMFuzzerTestOneInput(char *data, int size) {
+   // int length = strlen((const char *)data);
     string_modifier(data);
+    //printf(&res);
     return 0;
 }
